[refactor] Expression 분리에 따른 테스트 코드 위치 변경
